Eval/requirements.txt
absl-py==2.3.0
accelerate==1.10.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
beartype==0.14.1
beautifulsoup4==4.13.4
blake3==1.0.5
bleach==6.2.0
blessed==1.21.0
blis==1.3.0
blobfile==2.0.2
cachetools==6.0.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.9.4
confection==0.1.5
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
cymem==2.0.11
datasets==4.1.0
decorator==5.2.1
defusedxml==0.7.1
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.1
email_validator==2.2.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
executing==2.2.0
fancy-einsum==0.0.3
fastapi==0.115.12
fastapi-cli==0.0.7
fastchat==0.1.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
fire==0.7.0
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.3.0
gguf==0.17.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.7.0
interegular==0.3.3
ipython==8.12.3
jaxtyping==0.2.38
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lark==1.2.2
latex2sympy2_extended==1.10.2
litellm==1.73.2
llguidance==0.7.27
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru==0.7.3
lxml==4.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.6
mistune==3.1.3
modal==1.1.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.13
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.84.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-semantic-conventions-ai==0.4.9
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
plotext==5.3.2
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf>=5,<6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycountry==24.6.1
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.7
rouge_score==0.1.2
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
syllapy==0.7.2
sympy==1.14.0
synchronicity==0.10.2
tenacity==9.1.2
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==1.9.1
transformers==4.52.4
triton==3.3.0
typer==0.17.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vllm==0.9.0.1
wadler_lindig==0.1.3
wandb==0.19.7
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
word2number==1.1
wrapt==1.17.2
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zipp==3.22.0
